Add naive fallback for releases/index.yml files
Currently cloudfoundry/bosh has releases/bosh as a symlink to . which
causes `git cat-file`s to file. There does not appear to be a trivial
way to have git resolve symlinks are part of that lookup. A nicer
solution might be to remove the symlink from the bosh repository, but
that may take a while to get through. A better solution might be to
traverse symlinks internally since others might be using them for more
useful reasons.

For now, this change is being made to enable more adoption of this
resource. Can internally improve the file path resolution later.

Fixes https://github.com/dpb587/bosh-release-resource/issues/4